app/src/requirements.txt
Flask==3.0.0
gunicorn==22.0.0
Werkzeug==3.0.1
google-cloud-aiplatform==1.62.0
google-cloud-bigquery==3.21.0
google-cloud-storage==2.16.0
